Makefile: enable easier build on modern systems
Problem: Makefile assumes Lua includes in standard path and Lua library
is named liblua.so, but modern distros put includes in /usr/include/lua<ver>
and the library is liblua<ver>.

Update the makefile to allow LUA_VER to be set, and look for includes
in appropriate path and library with version name appended.

Build against Lua 5.3 by default.