lua: Update for Lua 5.2/5.3
Problem: The spank/lua plugin is not compatible with Lua 5.2, which
removed the LUA_GLOBALSINDEX special index. Also, per-coroutine globals
no longer seem to be supported, instead coroutines (as created by
luaL_newthread()) share global environment of the main Lua state,
so the hack to use these threads for each Lua script in order to get
a different per-script globals table no longer works.

Solution: Instead of using the hacky coroutine and shadow global table,
create a function environment (setfenv in 5.1, and _ENV/1st upvalue
in 5.2) for each compiled chunk (script), and save *that* table in
the registry, to later use to look up callbacks registered by the Lua
plugins. Dump the use of luaL_newthread(), it is no longer needed --
coroutines were not really being utlized as such anyway.